Remove status command, fold into get and logs
